Change JMC class inheritance structure, debugged Constants
