+++ src/jmc/cas/Evaluable.java
@@ -2,6 +2,7 @@ package jmc.cas;
 
 /**
  * Created by Jiachen on 20/05/2017.
+ * Evaluable
  */
 public interface Evaluable {
     double eval(double x);
