--- src/tests/FormatCoefficientTest.java
@@ -1,15 +1,0 @@
-package tests;
-
-/**
- * Created by student on 5/4/2017.
- */
-public class FormatCoefficientTest {
-    public static void main(String args[]){
-        //l(Function.formatCoefficients("3x+2x+3*x"));
-        //System.out.println("about".substring(1));
-    }
-
-    static void l(String s){
-        System.out.println(s);
-    }
-}
