+++ src/tests/GraphIndependenceTest.java
@@ -1,5 +1,6 @@
 package tests;
 
+import jmc.GraphFunction;
 import jmc.cas.Expression;
 import jmc.graph.Graph;
 import jui.JNode;
@@ -7,6 +8,7 @@ import processing.core.PApplet;
 
 /**
  * Created by Jiachen on 21/05/2017.
+ * Graph UI Test
  */
 public class GraphIndependenceTest extends PApplet {
     public static void main(String args[]) {
@@ -23,7 +25,7 @@ public class GraphIndependenceTest extends PApplet {
         JNode.init(this);
         JNode.add(new Graph(0, 0, width, height));
         Graph graph = ((Graph) JNode.getDisplayables().get(0));
-        graph.add(Expression.interpret("x"));
+        graph.add(new GraphFunction(Expression.interpret("x")));
 
     }
