+++ src/tests/JMathContextTest.java
@@ -5,7 +5,7 @@ import jmc.MathContext;
 import java.math.BigInteger;
 import java.util.ArrayList;
 
-public class BigIntegerFactorTest {
+public class JMathContextTest {
 
     public static void main(String[] args) {
         ArrayList<BigInteger> factors = MathContext.factor(new BigInteger(args[0]));
